fastapi-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
openai~=1.0
psycopg2-binary~=2.9
qdrant-client==1.6.0

SQLAlchemy~=2.0.23
passlib~=1.7.4
python-jose~=3.3.0

python-dotenv~=1.0.0
asyncpg